+++ twoliter/build.rs
@@ -40,6 +40,7 @@ fn main() {
     paths.copy_file("rpm2kmodkit");
     paths.copy_file("rpm2migrations");
     paths.copy_file("metadata.spec");
+    paths.copy_file("ocihelper");
 
     // Create tarball in memory.
     println!("Starting tarball creation at {:?}", SystemTime::now());
