+++ twoliter/embedded/build.Dockerfile
@@ -200,6 +200,11 @@ ARG KIT
 ARG PACKAGE_DEPENDENCIES
 ARG ARCH
 ARG NOCACHE
+ARG BUILD_ID
+ARG VERSION_ID
+ARG EXTERNAL_KIT_METADATA
+ARG VENDOR
+ARG LOCAL_KIT_DEPENDENCIES
 
 WORKDIR /home/builder
 USER builder
